fix model menu, update index and post templates

+++ menu/models.py
@@ -6,9 +6,9 @@ class Menu(models.Model):
     name = models.CharField(max_length=255, verbose_name='Название')
     description = models.TextField(blank=True, verbose_name='Описание')
     photo = models.ImageField(upload_to='photos/%Y/%m/%d/', verbose_name='Фото')
-    price = models.DecimalField(max_digits=9, decimal_places=2, verbose_name='Цена')
+    price = models.IntegerField(verbose_name='Цена')
     rating = models.DecimalField(max_digits=3, decimal_places=2, default=None, null=True, verbose_name='Рейтинг')
-    weight = models.DecimalField(max_digits=9, decimal_places=2, verbose_name='Граммовка')
+    weight = models.IntegerField(verbose_name='Граммовка')
     category = models.ForeignKey('Category', on_delete=models.PROTECT, verbose_name='Категория')
 
     def __str__(self):
